rename validation to formSchema, sanitizeString to toInputString
