+++ libs/react/shelter-operator/src/lib/components/beds/bed-form/constants/formSchema.ts
