+++ libs/react/shelter-operator/src/lib/components/beds/bed-form/formTypes.ts
@@ -1,4 +1,4 @@
 import type { z } from 'zod';
-import type { formSchema } from './constants/validation';
+import type { formSchema } from './constants/formSchema';
 
 export type BedFormData = z.infer<typeof formSchema>;
