+++ libs/react/shelter-operator/src/lib/components/reservations/reservation-form/formTypes.ts
@@ -1,4 +1,4 @@
 import type { z } from 'zod';
-import type { formSchema } from './constants/validation';
+import type { formSchema } from './constants/formSchema';
 
 export type ReservationFormData = z.infer<typeof formSchema>;
